exception added

+++ BlogApplication/src/main/java/com/application/blog/exception/ResourceNotFoundException.java
@@ -14,8 +14,8 @@ public class ResourceNotFoundException extends RuntimeException {
 	
 	private String resourceName;
 	private String feildName;
-	private Long feildValue;
-	public ResourceNotFoundException(String resourceName, String feildName, Long feildValue) {
+	private Integer feildValue;
+	public ResourceNotFoundException(String resourceName, String feildName, Integer  feildValue) {
 		super(resourceName+" not found with "+feildName+" "+feildValue);
 		this.resourceName = resourceName;
 		this.feildName = feildName;
